requirements.txt
streamlit
anthropic
openai
faiss-cpu
langchain
langchain-community
langchain-openai
python-dotenv
